fix: check postgres version for generated columns
Generated columns are only available on postgres version 12+

+++ src/driver/postgres/PostgresQueryRunner.ts
@@ -344,17 +344,21 @@ export class PostgresQueryRunner extends BaseQueryRunner implements QueryRunner 
                 return Promise.resolve();
             }));
 
-        // if table have column with generated type, we must add the expression to the meta table
-        await Promise.all(table.columns
-            .filter(column => column.generatedType === "STORED" && column.asExpression)
-            .map(async column => {
-                const tableName = await this.getTableNameWithSchema(table.name);
-                const deleteQuery = new Query(`DELETE FROM typeorm_generation_meta WHERE table_name = $1 AND column_name = $2`, [tableName, column.name]);
-                upQueries.push(deleteQuery);
-                upQueries.push(new Query(`INSERT INTO typeorm_generation_meta(table_name, column_name, generation_expression) VALUES ($1, $2, $3)`, [tableName, column.name, column.asExpression]));
-                downQueries.push(deleteQuery);
-                return Promise.resolve();
-            }));
+        if (await this.driver.isGeneratedColumnsSupported()) {
+            // if table have column with generated type, we must add the expression to the meta table
+            await Promise.all(table.columns
+                .filter(column => column.generatedType === "STORED" && column.asExpression)
+                .map(async column => {
+                    const tableName = await this.getTableNameWithSchema(table.name);
+                    const deleteQuery = new Query(`DELETE FROM typeorm_generation_meta WHERE table_name = $1 AND column_name = $2`, [tableName, column.name]);
+                    upQueries.push(deleteQuery);
+                    upQueries.push(new Query(`INSERT INTO typeorm_generation_meta(table_name, column_name, generation_expression) VALUES ($1, $2, $3)`, [tableName, column.name, column.asExpression]));
+                    downQueries.push(deleteQuery);
+                    return Promise.resolve();
+                }));
+        } else {
+            table.columns = table.columns.filter(column => !column.generatedType);
+        }
 
         upQueries.push(this.createTableSql(table, createForeignKeys));
         downQueries.push(this.dropTableSql(table));
@@ -575,7 +579,7 @@ export class PostgresQueryRunner extends BaseQueryRunner implements QueryRunner 
             downQueries.push(new Query(`ALTER TABLE ${this.escapePath(table)} DROP CONSTRAINT "${uniqueConstraint.name}"`));
         }
 
-        if (column.generatedType === "STORED" && column.asExpression) {
+        if (column.generatedType === "STORED" && column.asExpression && await this.driver.isGeneratedColumnsSupported()) {
             const tableName = await this.getTableNameWithSchema(table.name);
             const deleteQuery = new Query(`DELETE FROM typeorm_generation_meta WHERE table_name = $1 AND column_name = $2`, [tableName, column.name]);
             upQueries.push(deleteQuery);
@@ -906,7 +910,7 @@ export class PostgresQueryRunner extends BaseQueryRunner implements QueryRunner 
                 downQueries.push(new Query(`ALTER TABLE ${this.escapePath(table)} ALTER COLUMN "${newColumn.name}" TYPE ${this.driver.createFullType(oldColumn)}`));
             }
 
-            if (newColumn.generatedType !== oldColumn.generatedType) {
+            if (newColumn.generatedType !== oldColumn.generatedType && await this.driver.isGeneratedColumnsSupported()) {
                 // Convert generated column data to normal column
                 if (!newColumn.generatedType || newColumn.generatedType === "VIRTUAL") {
                     // We can copy the generated data to the new column
@@ -1010,7 +1014,7 @@ export class PostgresQueryRunner extends BaseQueryRunner implements QueryRunner 
             }
         }
 
-        if (column.generatedType === "STORED") {
+        if (column.generatedType === "STORED" && await this.driver.isGeneratedColumnsSupported()) {
             const tableName = await this.getTableNameWithSchema(table.name);
             upQueries.push(new Query(`DELETE FROM typeorm_generation_meta WHERE table_name = $1 AND column_name = $2`, [tableName, column.name]));
             downQueries.push(new Query(`INSERT INTO typeorm_generation_meta(table_name, column_name, generation_expression) VALUES ($1, $2, $3)`, [tableName, column.name, column.asExpression]));
@@ -1514,6 +1518,8 @@ export class PostgresQueryRunner extends BaseQueryRunner implements QueryRunner 
         if (!dbTables.length)
             return [];
 
+        const supportsGeneratedColumns = await this.driver.isGeneratedColumnsSupported();
+
         // create tables for loaded tables
         return Promise.all(dbTables.map(async dbTable => {
             const table = new Table();
@@ -1639,7 +1645,7 @@ export class PostgresQueryRunner extends BaseQueryRunner implements QueryRunner 
                         }
                     }
 
-                    if (dbColumn["is_generated"] === "ALWAYS" && dbColumn["generation_expression"]) {
+                    if (dbColumn["is_generated"] === "ALWAYS" && dbColumn["generation_expression"] && supportsGeneratedColumns) {
                         // In postgres there is no VIRTUAL generated column type
                         tableColumn.generatedType = "STORED";
                         // We cannot relay on information_schema.columns.generation_expression, because it is formatted different.
